Make the UX for the extension more similar to GitHub's
Instead of clicking the repo-stat item to toggle the star, it's now only
used as a display for the number of stars and the current status for the
user, more closely matching how the items next to it behave.

To toggle star/unstar, a button has been added next to Watch, Fork, etc.